Refactor:(#16) comment 반영
