+++ src/main/kotlin/team/sfe/server/global/redis/RedisConfig.kt
@@ -1,6 +1,5 @@
-package team.sfe.server.global.config
+package team.sfe.server.global.redis
 
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.connection.RedisConnectionFactory
@@ -8,14 +7,10 @@ import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactor
 
 @Configuration
 class RedisConfig(
-    @Value("\${spring.redis.host}")
-    val host: String,
-
-    @Value("\${spring.redis.host")
-    val port: Int
+    private val redisProperties: RedisProperties
 ) {
     @Bean
     fun redisConnectionFactory(): RedisConnectionFactory {
-        return LettuceConnectionFactory(host, port)
+        return LettuceConnectionFactory(redisProperties.host, redisProperties.port)
     }
 }
