+++ src/main/kotlin/team/sfe/server/global/redis/RedisProperties.kt
@@ -1,0 +1,9 @@
+package team.sfe.server.global.redis
+
+import org.springframework.boot.context.properties.ConfigurationProperties
+
+@ConfigurationProperties(prefix = "spring.redis")
+data class RedisProperties(
+    val host: String,
+    val port: Int
+)
